more specific check for lack of an incremental
Instead of catching any IOException for trying to fetch an incremental
and falling back to the full update package, check the respones code and
fall back if there's a 404 error. This avoids fallback when the issue is
caused by something else such as a failed connection.

This also closes the previous connection after getting the error so that
it can be reused for the new one via the HTTP connection pooling.