handle incremental being removed from server
If an incremental is removed from the server for some reason such as a
problem being identified with it, the update client should fall back to
using the full update instead. The case where this occurred in practice
was when we removed very old incrementals for end-of-life devices but a
few users were currently in the process of downloading them.